Stop adding link caption if it is null or agrees with the url
Formated links were of the form

  url (caption)

and Bugzilla typically autolinked the url. If caption equals with the
url, we ended up with comments as

  http://www.example.org/ (http://www.example.org/)

which is unnecessarily verbose. In such settings (and if no caption is
provided), we no condense the generated link to:

  http://www.example.org/

Bugzilla again autolinks the url, but the description that gave no
additional information is now gone.

Change-Id: I77543aa2c9d34dc4716d7e6e8176f379dafd747c