Adapt InitBugzilla init step to API change in the InitIts init step
With change [1] in plugins/its-base support for configuring branches
for the Bugzilla integration during init was added. Due to this change
the InitBugzilla init step needs to be adapted.

In addition update the example of the plugin initialisation in the
documentation because with change [1] there is now an additional
question during init for configuring the branches.

[1] https://gerrit-review.googlesource.com/53690

Change-Id: I66f349859cc6b228a308bc4ecf2ffa378a73571f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>