Drop failing tests
Ideally, we should fix the tests. We did for the newer stable branches.
But as gerrit upstream will soon stop support for stable-2.16, we
rather simply remove the failing tests. The plugin itself is working.
It's only the tests that did not get updated some time ago and are now
(obviously) failing. So they complain about no longer existing classes
(E.g.: ManiphestInfo) and PowerMock (which the ecosystem is moving away
from) is failing due to JDK/Library issues.

Change-Id: Ibfd4bee19926b46e492b4d81c2baa588c05153f7