Drop documentation for `remove-project`
`remove-project` has not fully been tested and was not doing what it
said on the tin. The newer stable branches have `remove-project` fixed.
But as gerrit upstream will soon end support for 2.16, we simple drop
support for the action instead of fixing on this old branch. That way,
old users of the plugin can still use it as they did before. And new
users of won't see the action and hence not use it.

Change-Id: I1f2ef0fcf57522b3d3c18fdb8cf4e4d6274f3ac6