Update patch set 1
Patch Set 1:

> Nope, i just presumed this file is used for the standalone mode.

Ok, I guess I should add some context.

The basic issue is that when building plugins like `its-phabricator`
you end up with a version like `4e23226` in the plugin list. That's
not overly useful. It tells you the commit it got built against, which
is great. But it does not tell you where the commit roughly
belongs. And the same commit can be built against multiple Gerrit
APIs. So which one got used for the build? Also, it does not mention
the `its-base` version it was built with.

We could cook all that info into the CI builds manually. But then
building locally brings back all of the above issues.

So we need to teach the build system to expose the API version and a
good version number.

Bringing the API version into the jar automatically is covered by
2d9a44d5f50696746b14df00025f68fb2613a92d.
Making the API version visible is in the `expose-api-version-of-plugins`topic [1]

Getting good version strings for plugins is at `plugin-version-stamps` [2].
In a first step, this topic improves on a plain `git describe` by adding more
heuristics. But regardless of how good the heuristics are, we'll for
sure find plugins where the heuristics are not good enough (like
`its-*` plugins). So it ultimately allows plugins to provide their own
`workspace_status.py` and thereby give plugins a way override their
version string.

This commit that we're at right now implements such a a
`workspace_status.py` for `its-phabricator` to provide a custom
version string (and re-uses the heuristics of Gerrit core for this.)

So this all will take us from before

  its-phabricator got built from commit 4e23226 (no clue about its-base or API)

to

  its-phabricator got built from commit 4e23226 on the stable-2.16
  branch with its-base commit 12345def on the stable-2.16 branch
  against Gerrit API 2.16.

. Having this additional information at hand is greatly useful when
finding out which code is currently running on a Gerrit server (Which
was one of the tricky parts at WMF's gerrit), and helps a lot when
trying to make sure the expected plugins with expected APIs run (which
has bitten us a few times in the past).


[1] https://gerrit-review.googlesource.com/q/topic:%22expose-api-version-of-plugins%22+(status:open%20OR%20status:merged)
[2] https://gerrit-review.googlesource.com/q/topic:%22plugin-version-stamps%22+(status:open%20OR%20status:merged)

Patch-set: 1
Reviewer: Gerrit User 1012155 <1012155@173816e5-2b9a-37c3-8a2e-48639d4f1153>